pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = "Python template with formater"
authors = ["Kazantsev Artem <me@lokach.ru>"]
readme = "README.md"
package-mode=false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.120.2"}
sqlalchemy = {extras = ["mypy"], version = "^2.0.44"}
alembic = "^1.17.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
python-multipart = "^0.0.20"
redis = "^7.0.1"
pydantic-settings = "^2.12.0"
python-dotenv = "^1.2.1"
pillow = "^12.0.0"
asyncpg = "^0.31.0"
pydantic = "^2.12.5"
aiogram = { version = "^3.23.0", python = "<3.15,>=3.10" }
gspread = "^6.2.1"
google-auth = "^2.43.0"
loguru = "^0.7.3"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pre-commit = "^4.0.1"
black = "^24.10.0"
ruff = "^0.8.3"
mypy = "^1.13.0"
locust = "^2.42.2"
pytest-asyncio = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 89
include = '\.pyi?$'

[tool.ruff]
lint.select = ["F", "E", "W", "I", "N", "PL", "C", "B"]
line-length = 89

[tool.ruff.lint.pep8-naming]
classmethod-decorators = ["classmethod", "pydantic.field_validator"]

[tool.pytest.ini_options]
pythonpath = [
    ".", "app"
]
